// src/types/setting.ts
export enum Language {
  eng = 'ENG',
  esp = 'ESP',
  cn = 'CN',
}

export enum TempScale {
  fahrenheit = 'FAHRENHEIT',
  celsius = 'CELSIUS',
}

export enum OTPStatus {
  none = 'none',
  request = 'request',
  verification = 'verification',
  verified = 'verified',
  notVerified = 'notVerified',
  error = 'error',
  resent = 'resent',
}